refactor API url
