+++ frontend/src/components/demos/SpeechASR.tsx
@@ -103,7 +103,7 @@ const stopRecording = () => {
       return;
     }
     setTableAIProgressLoading(true);
-    const serverEndpoint = serverBaseUrl + '/recipes/speech_asr_url/';
+    const serverEndpoint = serverBaseUrl + '/inference/speech_asr_url/';
   
     const formData = new FormData();
     formData.append('audio', audioFile);
