+++ frontend/src/components/indic_llm/IndicDemo.tsx
@@ -84,7 +84,7 @@ class IndicDemo extends Component<TextDemoProps, AppState> {
   sendPromptToServer = async () => {
     this.setState({tableAIProgressLoading:true});
 
-    const serverEndpoint = this.serverBaseUrl + '/recipes/indic_llm_url/';
+    const serverEndpoint = this.serverBaseUrl + '/inference/indic_llm_url/';
 
 
     const model = this.state.models.get(this.state.textSelectedModel);
